Fixed bug with exit/quit command.
When a command wasn't found, the child process wasn't killed. This meant that the *child* entered an infinite loop polling for user commands rather than the parent, meaning you'd have to type exit or quit N times, N being the number of children forked to run processes that couldn't be found.